Fix division bug and improve print function handling

+++ src/interpreter/cse_machine/__init__.py
@@ -257,7 +257,7 @@ class CSEMachine:
         "+": lambda rator, rand: rator + rand,
         "-": lambda rator, rand: rator - rand,
         "*": lambda rator, rand: rator * rand,
-        "/": lambda rator, rand: rator / rand,
+        "/": lambda rator, rand: int(rator / rand),
         "or": lambda rator, rand: rator or rand,
         "and": lambda rator, rand: rator and rand,
         "gr": lambda rator, rand: rator > rand,
